@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-mdl": "^1.11.0",
     "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.0"
