@@ -1,39 +1,50 @@
-import React, { Component } from 'react';
-import './App.css';
-import { Layout, Header, Navigation, Drawer, Content } from 'react-mdl';
-import Main from './components/main';
-import { Link } from 'react-router-dom';
+import React from "react";
+import "./App.css";
+import { Layout, Header, Navigation, Drawer, Content } from "react-mdl";
+import Home from "./pages/Home";
+import { Link } from "react-router-dom";
 
-class App extends Component {
-  render() {
-    return (
-      <div className="demo-big-content">
-    <Layout>
-        <Header className="header-color" title={<Link style={{textDecoration: 'none', color: 'white'}} to="/">MyPortfolio</Link>} scroll>
-            <Navigation>
-                <Link to="/resume">Resume</Link>
-                <Link to="/aboutme">About Me</Link>
-                <Link to="/projects">Projects</Link>
-                <Link to="/contact">Contact</Link>
-            </Navigation>
+const App = () => {
+  return (
+    <div className="demo-big-content">
+      <Layout>
+        <Header
+          className="header-color"
+          title={
+            <Link style={{ textDecoration: "none", color: "white" }} to="/">
+              MyPortfolio
+            </Link>
+          }
+          scroll
+        >
+          <Navigation>
+            <Link to="/resume">Resume</Link>
+            <Link to="/aboutme">About Me</Link>
+            <Link to="/projects">Projects</Link>
+            <Link to="/contact">Contact</Link>
+          </Navigation>
         </Header>
-        <Drawer title={<Link style={{textDecoration: 'none', color: 'black'}} to="/">MyPortfolio</Link>}>
-            <Navigation>
-              <Link to="/resume">Resume</Link>
-              <Link to="/aboutme">About Me</Link>
-              <Link to="/projects">Projects</Link>
-              <Link to="/contact">Contact</Link>
-            </Navigation>
+        <Drawer
+          title={
+            <Link style={{ textDecoration: "none", color: "black" }} to="/">
+              MyPortfolio
+            </Link>
+          }
+        >
+          <Navigation>
+            <Link to="/resume">Resume</Link>
+            <Link to="/aboutme">About Me</Link>
+            <Link to="/projects">Projects</Link>
+            <Link to="/contact">Contact</Link>
+          </Navigation>
         </Drawer>
         <Content>
-            <div className="page-content" />
-            <Main/>
+          <div className="page-content" />
+          <Home />
         </Content>
-    </Layout>
-</div>
-
-    );
-  }
-}
+      </Layout>
+    </div>
+  );
+};
 
 export default App;
